refactor(mobile): extract shared food-diary date-time helpers and drop unused spec import
